+++ src/github/auto-approve.ts
@@ -1,5 +1,5 @@
-import { Component } from '../../component';
-import { Project } from '../../project';
+import { Component } from '../component';
+import { Project } from '../project';
 
 /**
  * Options for 'AutoApprove'
