+++ src/github/index.ts
@@ -5,3 +5,6 @@ export * from './mergify';
 export * from './pr-template';
 export * from './mergify';
 export * from './workflows';
+export * from './auto-merge';
+export * from './auto-approve';
+export * from './auto-upgrade-dependencies';
